Add-GroupChats-Conditions-And-Error-handling

+++ Services/Groupchats/GroupChatsServices.cs
@@ -1,20 +1,63 @@
 ﻿using GData.DTOs.GroupchatsDTO;
 using GData.Entity;
+using GData.Exceptions;
 using GData.Repositories.GroupChat;
 using GData.Services.Users;
+using Microsoft.AspNetCore.Http.HttpResults;
+using Microsoft.EntityFrameworkCore;
 
 namespace GData.Services.Groupchats
 {
-    public class GroupChatsServices(IGroupChatRepository groupChatRepository, IAuthServices authServices) : IGroupChatsServices
+    public class GroupChatsServices(IGroupChatRepository groupChatRepository, IAuthServices authServices, GroupChatExceptionList groupChatExceptionList) : IGroupChatsServices
     {
         public async Task<Groupchat> AddUserToGroupChatService(Guid creatorId, Guid userId, Guid Id)
         {
             
-           var user= await authServices.GetUserByIdService(userId);
-           var groupChat = await GetGroupChatByIdService(Id);
-            
-           await groupChatRepository.AddUserToGroupChat(user, groupChat);
-           return groupChat;
+            var user= await authServices.GetUserByIdService(userId);
+            var groupChat = await GetGroupChatByIdService(Id);
+            var creator = await authServices.GetUserByIdService(creatorId);
+
+            if (groupChat is null)
+            {
+
+                return await groupChatExceptionList.EditGroupChatChatDoesNotExist();
+
+            }
+
+            if (groupChat.CreatorId != creatorId)
+            {
+
+                return await groupChatExceptionList.CreatorNotValid();
+
+            }
+
+            if (creator is null)
+            {
+
+                return await groupChatExceptionList.EditGroupChatChatDoesNotExist();
+
+            }
+
+            if (creator.IsEmailConfirmed == false)
+            {
+
+                return await groupChatExceptionList.UnverifiedUserEmail();
+
+            }
+
+            foreach (var member in groupChat.ChatMembers)
+            {
+
+                if (member.Id == userId)
+                {
+
+                    return await groupChatExceptionList.AddUserAlreadyAdded();
+
+                }
+            }
+
+            await groupChatRepository.AddUserToGroupChat(user, groupChat);
+            return groupChat;
 
         }
 
@@ -22,6 +65,34 @@ namespace GData.Services.Groupchats
         {
 
             var creator=await authServices.GetUserByIdService(creatorId);
+
+            if (creator is null)
+            {
+
+                return await groupChatExceptionList.ChatCreatorDoesNotExist();
+
+            }
+
+            if (creator.IsEmailConfirmed == false)
+            {
+
+                return await groupChatExceptionList.UnverifiedUserEmail();
+
+            }
+
+            if (string.IsNullOrWhiteSpace(request.ChatTitle))
+            {
+
+                return await groupChatExceptionList.ChatNameNeedsToHaveData();
+
+            }
+
+            if(request.ChatTitle.Length<3)
+            {
+
+                return await groupChatExceptionList.ChatNameNeedsToHaveMoreThanThreeChars();
+
+            }
 
             var groupchat = new Groupchat()
             {
@@ -34,6 +105,14 @@ namespace GData.Services.Groupchats
             await groupChatRepository.CreateGroupChat(groupchat);
 
             var createdGroupChat=await GetGroupChatByChatNameService(request.ChatTitle);
+
+            if(createdGroupChat is null)
+            {
+
+                return await groupChatExceptionList.EditGroupChatChatDoesNotExist();
+
+            }
+
             await AddUserToGroupChatService(creatorId,creatorId, createdGroupChat.Id);
             return groupchat;
 
@@ -44,6 +123,36 @@ namespace GData.Services.Groupchats
             
             var groupChat=await groupChatRepository.GetGroupchatById(Id);
 
+            var creator = await authServices.GetUserByIdService(creatorId);
+
+            if (groupChat is null)
+            {
+
+                return await groupChatExceptionList.EditGroupChatChatDoesNotExist();
+
+            }
+
+            if (groupChat.CreatorId != creatorId)
+            {
+
+                return await groupChatExceptionList.CreatorNotValid();
+
+            }
+
+            if (creator is null)
+            {
+
+                return await groupChatExceptionList.EditGroupChatChatDoesNotExist();
+
+            }
+
+            if (creator.IsEmailConfirmed == false)
+            {
+
+                return await groupChatExceptionList.UnverifiedUserEmail();
+
+            }
+
             await groupChatRepository.DeleteGroupChat(groupChat);
             return groupChat;
 
@@ -53,6 +162,50 @@ namespace GData.Services.Groupchats
         {
 
             var groupChat = await groupChatRepository.GetGroupchatById(Id);
+
+            var creator = await authServices.GetUserByIdService(creatorId);
+
+            if(groupChat is null)
+            {
+
+                return await groupChatExceptionList.EditGroupChatChatDoesNotExist();
+
+            }
+            
+            if(groupChat.CreatorId != creatorId)
+            {
+
+                return await groupChatExceptionList.CreatorNotValid();
+
+            }
+            
+            if(creator is null)
+            {
+
+                return await groupChatExceptionList.EditGroupChatChatDoesNotExist();
+
+            }
+            
+            if (creator.IsEmailConfirmed == false)
+            {
+
+                return await groupChatExceptionList.UnverifiedUserEmail();
+
+            }
+            
+            if (string.IsNullOrWhiteSpace(request.ChatTitle))
+            {
+
+                return await groupChatExceptionList.ChatNameNeedsToHaveData();
+
+            }
+
+            if(request.ChatTitle.Length<3)
+            {
+
+                return await groupChatExceptionList.ChatNameNeedsToHaveMoreThanThreeChars();
+
+            }
 
             await groupChatRepository.EditGroupChatTitle(groupChat,request);
             return groupChat;
@@ -70,6 +223,14 @@ namespace GData.Services.Groupchats
         {
             
             var groupChat=await groupChatRepository.GetGroupchatById(Id);
+
+            if(groupChat is null)
+            {
+
+                return await groupChatExceptionList.EditGroupChatChatDoesNotExist();
+
+            }
+
             return groupChat;
 
         }
@@ -78,6 +239,14 @@ namespace GData.Services.Groupchats
         {
             
             var groupChat=await groupChatRepository.GetGroupchatByChatName(title);
+
+            if (groupChat is null)
+            {
+
+                return await groupChatExceptionList.EditGroupChatChatDoesNotExist();
+
+            }
+
             return groupChat;
 
         }
@@ -85,8 +254,51 @@ namespace GData.Services.Groupchats
         public async Task<Groupchat> RemoveUserFromGroupChatService(Guid creatorId, Guid userId, Guid Id)
         {
 
+            var creator = await authServices.GetUserByIdService(creatorId);
             var user = await authServices.GetUserByIdService(userId);
             var groupChat = await GetGroupChatByIdService(Id);
+
+            if (groupChat is null)
+            {
+
+                return await groupChatExceptionList.EditGroupChatChatDoesNotExist();
+
+            }
+
+            if (groupChat.CreatorId != creatorId)
+            {
+
+                return await groupChatExceptionList.CreatorNotValid();
+
+            }
+
+            if (creator is null)
+            {
+
+                return await groupChatExceptionList.EditGroupChatChatDoesNotExist();
+
+            }
+
+            if (creator.IsEmailConfirmed == false)
+            {
+
+                return await groupChatExceptionList.UnverifiedUserEmail();
+
+            }
+
+            if(user==creator)
+            {
+
+                return await groupChatExceptionList.RemoveUserIsOwner();
+
+            }
+
+            if(user is null)
+            {
+
+                return await groupChatExceptionList.UserAlreadyRemoved();
+
+            }
 
             await groupChatRepository.RemoveUserFromGroupChat(user, groupChat);
             return groupChat;
